lib.systems.flakeExposed: Rely on lib.systems.supported
`lib.systems.supported` is intended to be the authoritative source on
what support tier platforms belong to. By getting the platforms built by
Hydra from there we make explicit which platforms are currently not
supported according to their tier and should either be taken care of or
reassigned to a different tier.

`x86_64-unknown-linux-musl` is listed in RFC #46 as a tier 3 platform
but is not currently built by Hydra for example. And many platforms
originally proposed as tier 7 are in fact built by Hydra and at least
some of them should be considered for promotion.