nixos/ssh: avoid IFD when using programs.ssh.knowHostsFiles
When passing a derivation, the call to copyPathToStore will look at the
contents of the outPath and copy the whole file or directory to a new
store path. Because we look at the contents of the derivation, we trigger
an IFD.

To avoid this, we deprecate that option and introduce an alternative
one, knowHostsPaths, that takes a store path and simply coerces the
given entries into a string. Nix will then correctly realise that
if we pass it a derivation, it can just substitute it with the derivation's
outPath.
Actual path entries will still be copied to the store as before.