mpv: use the Swift stdenv on Darwin
This fixes an issue where headers from the stdenv’s clang are being used
by Swift’s clang. That does not cause issues when the stdenv on Darwin
uses clang 11, but the build will fail on aarch64-darwin with a stdenv
that uses clang 16 (due to changes in `arm_neon.h` that are not
compatible with clang 15).