CODEOWNERS: add ajoseph for kernel/manual-config.nix
It is hard to get people to test changes to our kernel expression on
anything other than x86_64 and aarch64 before merging changes.

Therefore, I would like to be notified of PRs opened agains this
file, so that I can hopefully catch breakage before it is merged,
rather than afterwards.