all-cabal-hashes: 2023-07-19T20:56:38Z -> 2023-07-24T19:28:29Z
The main motivation for this is that the latest versions of hspec-core
and hspec-expectations got out of sync due to an unlucky timing on the
hackage snapshot update. As a consequence, we weren't able to build
cabal-install in some package sets. Additionally, this brings a version
of futhark that can be built with the lsp version we ship.

This commit has been generated by maintainers/scripts/haskell/update-hackage.sh
and maintainers/scripts/haskell/regenerate-hackage-packages.nix.

Additional changes:

* Adapt to new xhtml version (still doesn't match the version normally
  shipped by the respective GHC as a core library).

* Adapt to new versions of hspec* and pandoc