expand-response-params: explain what a "response file" is
I hadn't heard of this term before, so I assume other people might
not have either.  This commit adds a comment to
expand-response-params explaining what a "response file" is.

Most mentions of "response file" in Nixpkgs are near some kind of
reference to this derivation, so anybody trying to figure out what
this is will encounter this comment pretty quickly.