stdenv: add `stdenv.isAvailable pkg`
This commit adds `stdenv.isAvailable pkg`, which checks `meta`
availability not only for the hostPlatform, but also for the
`buildPlatform` and `targetPlatform` using
`lib.meta.checkAvailability`.

Our current meta.{platforms,badPlatforms} scheme isn't capable of
expressing constraints that apply to anything other than the
hostPlatform.  For example:

- LLVM can't be used with targetPlatform.isMips64n32, but GCC can

The clumsy way of working around this is to have `buildRustPackage`
inject `meta.badPlatforms` values into the packages it creates, but
that covers only Rust (not other LLVM-based compilers) and only
those which use `buildRustPackage`.  Really, it's the wrong layer at
which to express these things.

This commit adds `meta.availability.{build,host,target}.{bad,only}`
which are computed automatically based on a package's dependencies.

This is done in a backwards-compatible way for the hostPlatform:
`meta.availability.host` will incorporate
`meta.{platforms,badPlatforms}` from both the package as well as all
of its `buildInputs`.  The constraints are cumulative, so deep
recursion (like checkMetaRecursively) is not needed; each package's
`meta.availability` only needs to look at that package's immediate
inputs.

This commit also causes check-meta.nix to utilize the information in
meta.badBuildPlatforms and meta.badTargetPlatforms.

In particular, the following test commands now behave correctly:

  # XFAIL: golang can't build *on* MIPS, but it can cross compile to it
  nix-instantiate . -A pkgsCross.mipsel-linux-gnu.go

  # SUCCESS: golang can't build *on* MIPS, but it can cross compile to it
  nix-instantiate . -A pkgsCross.mipsel-linux-gnu.dnscrypt-proxy2

  # XFAIL: golang can't target Mips64n32
  nix-instantiate . -A pkgsCross.mips64el-linux-gnuabin32.dnscrypt-proxy2

  # XFAIL: rust can't target Mips64n32
  nix-instantiate . -A pkgsCross.mips64el-linux-gnuabin32.tiny