doc: minimal documentation of supported platforms
This commit adds minimal documentation of the supported platforms.
More exhaustive documentation would require producing a list of
platforms for each of the 7 tiers. This was attempted in #245368, but it
quickly became clear that that would be a long-term effort.

In the meantime, this commit adds the most important information to the
manual.

Co-authored-by: Valentin Gagarin <valentin.gagarin@tweag.io>