extractDarwinApp: support more source formats
The arguments and documentation implied that only `.app` and `.dmg`
we supported, but that wasn't true.
The case for `src = ./some/folder` and non-dmg archives with multiple
folders was not supported.

Now, the user can specify srcHasManySubfolders which sets the
`sourceRoot` for `mkDerivation` to a default value.
The user also can directly set `sourceRoot` themselves.