amend! python310Packages.bokeh: 2.4.3 -> 3.2.0
python310Packages.bokeh: 2.4.3 -> 3.2.1

https://docs.bokeh.org/en/latest/docs/releases.html#release-3-2-1

Changes the build to a source build and enables tests.

To enable the tests we have to fetch the sources from Git, which in turn
requires us to build the static files from source.

Bokeh uses setuptools-git-versioning, which has no version override, so
we also fix the broken fallback version in the package metadata.

We skip sampledata tests, because that would require an additional
download, and we skip selenium tests because they are crashing with
chromium/chromedriver and getting out of sync with geckodriver/firefox.