ldgallery: drop package
This removes 'ldgallery' from nixpkgs.
It remains available through the Nix Flake provided by ldgallery.

The package of its "compiler" part was already marked as broken due to
dependencies having gone out of sync with the rest of the Haskell
packages.

The package of its "viewer" part also got out of sync due to changes
in JS dependency manager.