gdb: disable sim by default
Out of the box, building GDB will also build the GNU simulator
project[1].  When building GDB with --enable-target=all, this result in
all simulators to be built, which accounts for a significant amount of
space.

This patch makes the default GDB package to not bundle sim.  Sim can be
enabled using `enableSim = true`;

As a result, the default GDB install is now about 40M v.s. 650M when all
simulators are included.

[1] https://sourceware.org/gdb/wiki/Sim