libreoffice-qt: fix build
Broke in 408ece7d3d5b68f5c0d870abb936ee8950a1b9f9 because the
`disallowedRequisites` fails here since the QT variant apparently needs
to reference a few dev outputs[1].

I won't look into the details of that now, so the easiest way to unbreak
is to skip the check for the QT variant. It should be kept for non-QT
though to make sure that a change similar to the BUILDCONFIG thing isn't
missed again by us.

[1] https://github.com/NixOS/nixpkgs/pull/245361#issuecomment-1651389735

    error: output '/nix/store/2y0czyy26gcsqhmcvd2mlqa35f0gcl4l-libreoffice-7.5.4.1' is not allowed to refer to the following paths:
             /nix/store/0hmvklj0mbhrn8flwbcwivvkv45limhg-freetype-2.13.0-dev
             /nix/store/0rnx7rc87hwkbrhsys7mgwq4jw6pz7ma-zlib-1.2.13-dev
             [...]