lib.fixedPoints.extends: document order of invocation
The description for lib.fixedPoints.extends is very hard to
understand.  I read it a few times trying to figure out "in what
order are its arguments applied by the resulting function?" and
couldn't figure that out.  So I added the answer to that question in
a comment.

I don't think the allusion to Java is particularly helpful.