nixos/kernel: Added optional support for consoleLogLevel
Ran into an issue on my device where specifying a loglevel in the kernel parameters prevents it from entering a multi-user session properly. This change allows the loglevel parameter to be omitted from the options list. 

Tested against version 23.05.git.48e82fe1b1cM (Stoat) and verified the option is removed when null is specified. Very rusty to nix so my patch may not be the most ideal. Contributions are encouraged.