yambar: fix strictDeps compilation
Yambar uses meson as build system. Its `dependency` function uses pkg-config as
default method for finding the required dependencies.

However, as seen in their code[^1], scdoc is searched with `native` set to
true. According to Meson specs[^2], it directs the search to the _build_ machine
instead of the _host_ one. This parameter is usually needed only when finding a
tool to be used during compilation.

Because of this, `pkg-config` needs to be included at `depsBuildBuild`.

Why? Magic! Sorry, I am not so proficient in cross-compilation issues yet.

[^1] https://codeberg.org/dnkl/yambar/src/commit/78f7b60e1308d9f6c879b5cbae2dadba4fa88762/doc/meson.build#L3
[^2] https://mesonbuild.com/Reference-manual_functions.html#dependency_native