srt-to-vtt-cl: fix macOS builds
The upstream Makefile is using logic to build into different dirs
depending on uname output. Trivial to get macOS builds working if
we just don't do that.

I've also stripped out a use of substituteAll that was having no
effect (the variables it replaced were removed during review of
the initial PR.)