lib.systems.examples: use four component triples for embedded
Thanks to a [change in gnu-config] we no longer have to rely on
gnu-config interpreting `none` as a vendor and nixpkgs as a kernel.
Instead we can specify a four component triple that corresponds to how
nixpkgs interprets the shortened one because gnu-config now accepts
them (the long form at least).

As a result, we also need to make sure that an up to date gnu-config is
used for the affected platforms (everything with kernel none).

Reference #165836.

[change in gnu-config]: https://git.savannah.gnu.org/cgit/config.git/commit/?id=998ba1414387b4ce1a519be234e1609bc7912e0c