nixos/filesystems/zfs: load udev rules in stage 1 too
Previously, we were not loading udev rules in stage 1, making
/dev/zfs 0600 by virtue of "default behavior", the problem with this
is that it makes it impossible to use ZFS tooling as a user for no real reason.

Moreover, in systemd pre-v254, the udev rules was probably applied again in stage 2
and made the wrong implementation invisible because no one was running zfsutil
as non-root that early.

We load the udev rules in stage 1 now, too, which is necessary in systemd v254 and more,
otherwise /dev/zfs will be left 0600 instead of 0666.