libgpiod: fix Python bindings
libgpiod's Python bindings no longer worked after the upgrade to 2.0.1. The
build system installs an egg, which doesn't work in nixpkgs. To fix this, I
adopted the same approach I took in #204884. This patch builds the Python
bindings as a separate package, using the normal nixpkgs Python infrastructure.
Besides fixing the bindings, this has the added benefit of avoiding the need to
build a redundant copy of libgpiod as part of the Python bindings package.

Lastly, I cleaned up the libgpiod package a bit, removing an unused dependency
on kmod and an unnecessary configure flag. I also added the full list of
licenses that apply to the package.