nixos/nextcloud: improve documentation of `adminuser`-option
This option only has an effect on the initial setup of Nextcloud and
changes later won't have any effect.

Same issue as with `adminpassFile` - it's only passed to the installer
command - but even worse because the username is frequently used as
unique ID in the database, so there's no trivial way to change it, even
imperatively.

(cherry picked from commit 648546b09dffcaed6a1a6018e783380e169ab9c3)