nixosTests.caddy: remove etag subtest
Caddy 2.7.x does no longer return etags for files with unix modtimes of
0 and 1.
Files in /nix/store have a modtime of 1.

This is something that has been specifically implemented for nix.

For now, we decided to remove the test.
But I might reimplement a similar etag subtest some time in the future.