qt6.qtbase: fix macdeployqt would not find qmlimportscanner
The qmlimportscanner tool is provided by qtdeclarative. Because of the
modularized installation in Nix, it can not be found via the usual
mechanisms.  Also, hard-coding it like we do for Qt5 would also not
work, as it would require making qtbase depend on qtdeclarative.

Here we add an option to provide its location via the environment.
While this means macdeployqt does not work out of the box, it provides
a workaround for users.

Also, we make sure that qmlimportscanner gets passed the right QML
import paths as described in the environment.