musescore: re-enable using system freetype
Fixes #244409.
Upstream removed the option to use system freetype library in v4.1.0,
causing the app to crash on systems when the outdated bundled freetype
tries to load the Noto Sans font.
For now, re-add the option ourselves. The fix has been merged upstream,
so we can remove this patch with the next version. In the future, we
may replace the other bundled thirdparty libs with system libs.