gdal: introduce 'useMinimalFeatures' flag to reduce closure size
Summary: GDAL has an atrociously large closure weighing in at over
1.3GiB. This makes many things such as packaging PostGIS/PostgreSQL
inside Docker images, or even copying closures, perform pretty poorly.

More importantly, adding unlimited surface area to GDAL is effectively
a support and maintenance burden. This is one of the major motivators
of this patch: to make it possible to reduce the number of supported
raster formats and thus scope support and maintence to specific subsets
of functionality.

Signed-off-by: Austin Seipp <aseipp@pobox.com>