postgresql_beta: init at version 16beta3
Summary: For integration and testing purposes, especially for downstream
consumers (Flakes, etc) that rely on testing against various PostgreSQL
versions, it's useful to have a beta version around. An example of this
is PostgREST, which has several people using Nix in order to test the
server against different versions. Testing against beta versions is
currently awkward because there's no way to get them besides patching
and forking Nixpkgs.

There's a related problem here, which is that the APIs needed to create
your *own* postgresql package, with all extensions, are not exposed.
That might be worth doing later on; but for now, it's easy to just track
beta versions and nip the issue in the bud from there.

The intent is that this `beta` package follows the same basic rules as
the Linux prerelease "beta" packages; that is:

- `postgresql_beta` *always* tracks a beta, even if the corresponding
  release is out.
   - That means that once PostgreSQL 16 is released, this expression
     will be out of date until PostgreSQL 17 Beta 1 is released.
- `postgresql_beta` is updated on a provisional basis i.e. it won't
   necessarily get prompt updates or security updates like the other
   expressions.
- It is primarily intended for advanced users to test and integrate
  against, so they might submit issues to various upstream maintainers.

Many thanks to Laurence Isla, who initially made this change as part of
an effort to integrate PostgREST against PostgreSQL 16 Beta 2.

See also: postgrest/postgrest#2865

Co-authored-by: Laurence Isla <lau.isla.c@gmail.com>
Signed-off-by: Austin Seipp <aseipp@pobox.com>