nextcloudPackages*: pin `maps` to stable 1.1.0
No idea what this gibberish disguised as tag is, but nc4nix doesn't seem
to cope well with it. For now, let's pin `maps` to the stable 1.1 release (as
it's the case for nextcloud27 already) since 1.1 is supported for all of
v25 to v27[1], so this seems reasonable to do.

[1] https://github.com/nextcloud/maps/blob/v1.1.0/appinfo/info.xml#L36